InterestRateField fixes & improvements
- In the env vars, add an auto delegate per branch
- Enable the IC strategies
- Fetch the min & max interest rates from the contracts
- Split delegation modal components in separate files
- Start migrating some subgraph-hooks functions into liquity-utils.ts
