+++ frontend/app/src/comps/InterestRateField/InterestRateField.tsx
@@ -1,35 +1,25 @@
 import type { Address, CollIndex, Delegate } from "@/src/types";
 import type { Dnum } from "dnum";
-import type { ReactNode } from "react";
 
-import { Amount } from "@/src/comps/Amount/Amount";
 import { INTEREST_RATE_DEFAULT, INTEREST_RATE_MAX, INTEREST_RATE_MIN } from "@/src/constants";
 import content from "@/src/content";
 import { IC_STRATEGIES } from "@/src/demo-mode";
 import { jsonStringifyWithDnum } from "@/src/dnum-utils";
 import { useInputFieldValue } from "@/src/form-utils";
-import { fmtnum, formatRedemptionRisk } from "@/src/formatting";
-import { getRedemptionRisk } from "@/src/liquity-math";
+import { fmtnum } from "@/src/formatting";
 import { useInterestRateChartData } from "@/src/liquity-utils";
-import { useInterestBatchDelegate } from "@/src/subgraph-hooks";
-import { infoTooltipProps, riskLevelToStatusMode } from "@/src/uikit-utils";
+import { infoTooltipProps } from "@/src/uikit-utils";
 import { noop } from "@/src/utils";
 import { css } from "@/styled-system/css";
 import {
-  AddressField,
-  AnchorTextButton,
-  Button,
   Dropdown,
   HFlex,
-  IconCopy,
   InfoTooltip,
   InputField,
   lerp,
-  Modal,
   norm,
   shortenAddress,
   Slider,
-  StatusDot,
   TextButton,
 } from "@liquity2/uikit";
 import { blo } from "blo";
@@ -37,15 +27,16 @@ import * as dn from "dnum";
 import Image from "next/image";
 import { memo, useId, useState } from "react";
 import { match } from "ts-pattern";
-
-import icLogo from "./ic-logo.svg";
+import { DelegateModal } from "./DelegateModal";
+import { IcStrategiesModal } from "./IcStrategiesModal";
+import { MiniChart } from "./MiniChart";
 
 export type DelegateMode = "manual" | "strategy" | "delegate";
 
 const DELEGATE_MODES: DelegateMode[] = [
   "manual",
   "delegate",
-  // "strategy",
+  "strategy",
 ];
 
 export const InterestRateField = memo(
@@ -92,8 +83,12 @@ export const InterestRateField = memo(
 
     const interestChartData = useInterestRateChartData(collIndex);
 
-    const interestRateNumber = interestRate && dn.toNumber(dn.mul(interestRate, 100));
-    const chartdataPoint = interestChartData.data?.find(({ rate }) => rate === interestRateNumber);
+    const interestRateNumber = interestRate && dn.toNumber(
+      dn.mul(interestRate, 100),
+    );
+    const chartdataPoint = interestChartData.data?.find(
+      ({ rate }) => rate === interestRateNumber,
+    );
     const boldRedeemableInFront = chartdataPoint?.debtInFront ?? dn.from(0, 18);
 
     const handleDelegateSelect = (delegate: Delegate) => {
@@ -122,7 +117,9 @@ export const InterestRateField = memo(
                 <Slider
                   gradient={[1 / 3, 2 / 3]}
                   gradientMode="high-to-low"
-                  chart={interestChartData.data?.map(({ size }) => Math.max(0.1, size)) ?? []}
+                  chart={interestChartData.data?.map(
+                    ({ size }) => Math.max(0.1, size),
+                  ) ?? []}
                   onChange={(value) => {
                     fieldValue.setValue(String(
                       Math.round(
@@ -146,7 +143,9 @@ export const InterestRateField = memo(
               <TextButton
                 size="large"
                 label={delegate
-                  ? IC_STRATEGIES.find(({ address }) => address === delegate)?.name
+                  ? IC_STRATEGIES.find(
+                    ({ address }) => address === delegate,
+                  )?.name
                   : "Choose strategy"}
                 onClick={() => {
                   setDelegatePicker("strategy");
@@ -192,15 +191,9 @@ export const InterestRateField = memo(
             end: (
               <div>
                 <Dropdown
-                  items={DELEGATE_MODES.map((mode) => {
-                    const modeContent = content.interestRateField.delegateModes[mode];
-                    return ({
-                      label: modeContent.label,
-                      secondary: modeContent.secondary,
-                      disabled: mode === "strategy",
-                      disabledReason: "Coming soon",
-                    });
-                  })}
+                  items={DELEGATE_MODES.map((mode) => (
+                    content.interestRateField.delegateModes[mode]
+                  ))}
                   menuWidth={300}
                   menuPlacement="end"
                   onSelect={(index) => {
@@ -225,12 +218,16 @@ export const InterestRateField = memo(
                     ? fmtnum(boldInterestPerYear)
                     : "−"} BOLD / year
                 </div>
-                <InfoTooltip {...infoTooltipProps(content.generalInfotooltips.interestRateBoldPerYear)} />
+                <InfoTooltip
+                  {...infoTooltipProps(
+                    content.generalInfotooltips.interestRateBoldPerYear,
+                  )}
+                />
               </HFlex>
             ),
             end: (
               <span>
-                <span>{"Redeemable before you: "}</span>
+                Redeemable before you:{" "}
                 <span
                   className={css({
                     fontVariantNumeric: "tabular-nums",
@@ -288,53 +285,22 @@ export const InterestRateField = memo(
               : null
           }
         />
-        <Modal
+        <DelegateModal
+          collIndex={collIndex}
           onClose={() => {
             setDelegatePicker(null);
           }}
-          title={
-            <div
-              className={css({
-                display: "flex",
-                alignItems: "center",
-                gap: 10,
-              })}
-            >
-              <div>
-                {content.interestRateField.icStrategyModal.title}
-              </div>
-              <Image
-                alt=""
-                src={icLogo}
-                width={24}
-                height={24}
-              />
-            </div>
-          }
-          visible={delegatePicker === "strategy"}
-        >
-          <DelegatesModalContent
-            collIndex={collIndex}
-            chooseLabel="Choose"
-            delegates={IC_STRATEGIES}
-            intro={content.interestRateField.icStrategyModal.intro}
-            onSelectDelegate={handleDelegateSelect}
-          />
-        </Modal>
-        <Modal
-          onClose={() => {
-            setDelegatePicker(null);
-          }}
-          title={content.interestRateField.delegatesModal.title}
+          onSelectDelegate={handleDelegateSelect}
           visible={delegatePicker === "delegate"}
-        >
-          <CustomDelegateModalContent
-            collIndex={collIndex}
-            chooseLabel="Set delegate"
-            intro={content.interestRateField.delegatesModal.intro}
-            onSelectDelegate={handleDelegateSelect}
-          />
-        </Modal>
+        />
+        <IcStrategiesModal
+          collIndex={collIndex}
+          onClose={() => {
+            setDelegatePicker(null);
+          }}
+          onSelectDelegate={handleDelegateSelect}
+          visible={delegatePicker === "strategy"}
+        />
       </>
     );
   },
@@ -342,438 +308,3 @@ export const InterestRateField = memo(
     jsonStringifyWithDnum(prev) === jsonStringifyWithDnum(next)
   ),
 );
-
-function CustomDelegateModalContent({
-  collIndex,
-  intro,
-  onSelectDelegate,
-}: {
-  collIndex: CollIndex;
-  chooseLabel: string;
-  intro: ReactNode;
-  onSelectDelegate: (delegate: Delegate) => void;
-}) {
-  const [delegateAddress, setDelegateAddress] = useState<null | Address>(null);
-  const [delegateAddressValue, setDelegateAddressValue] = useState("");
-
-  const delegate = useInterestBatchDelegate(collIndex, delegateAddress);
-
-  return (
-    <>
-      <div
-        className={css({
-          fontSize: 16,
-          color: "contentAlt",
-        })}
-      >
-        {intro}
-      </div>
-      <div
-        className={css({
-          paddingTop: 40,
-        })}
-      >
-        <form
-          onSubmit={(event) => {
-            event.preventDefault();
-            if (delegate.data) {
-              onSelectDelegate(delegate.data);
-            }
-          }}
-        >
-          <AddressField
-            onAddressChange={setDelegateAddress}
-            onChange={setDelegateAddressValue}
-            placeholder="Enter delegate address"
-            value={delegateAddressValue}
-          />
-        </form>
-      </div>
-
-      <div
-        className={css({
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          gap: 8,
-          paddingTop: 32,
-          paddingBottom: 24,
-          minHeight: 312,
-        })}
-      >
-        {delegateAddress
-          ? (
-            <div
-              className={css({
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "center",
-                width: "100%",
-                height: "100%",
-              })}
-            >
-              {delegate.status === "pending"
-                ? (
-                  <div
-                    className={css({
-                      color: "contentAlt",
-                      paddingTop: 40,
-                    })}
-                  >
-                    Loading…
-                  </div>
-                )
-                : delegate.status === "error"
-                ? (
-                  <div
-                    className={css({
-                      color: "contentAlt",
-                      paddingTop: 40,
-                    })}
-                  >
-                    Error: {delegate.error?.name}
-                  </div>
-                )
-                : (
-                  delegate.data
-                    ? (
-                      <DelegateBox
-                        delegate={delegate.data}
-                        selectLabel="Choose"
-                        onSelect={onSelectDelegate}
-                      />
-                    )
-                    : (
-                      <div>
-                        The address is not a valid{" "}
-                        <AnchorTextButton
-                          label="delegate"
-                          href="https://docs.liquity.org/v2-faq/redemptions-and-delegation#what-is-delegation-of-interest-rates"
-                          external
-                        />.
-                      </div>
-                    )
-                )}
-            </div>
-          )
-          : (
-            <>
-              <div>
-                Set a valid{" "}
-                <AnchorTextButton
-                  label="delegate"
-                  href="https://docs.liquity.org/v2-faq/redemptions-and-delegation#what-is-delegation-of-interest-rates"
-                  external
-                />{" "}
-                address.
-              </div>
-
-              <div>
-                Delegate addresses can be found{"  "}
-                <AnchorTextButton
-                  label="here"
-                  href="https://docs.liquity.org/v2-faq/redemptions-and-delegation#docs-internal-guid-441d8c3f-7fff-4efa-6319-4ba00d908597"
-                  external
-                />.
-              </div>
-            </>
-          )}
-      </div>
-    </>
-  );
-}
-
-function DelegatesModalContent({
-  collIndex: _collIndex,
-  delegates = [],
-  chooseLabel: _chooseLabel,
-  intro,
-  onSelectDelegate: _onSelectDelegate,
-}: {
-  collIndex: CollIndex;
-  delegates?: Delegate[];
-  chooseLabel: string;
-  intro: ReactNode;
-  onSelectDelegate: (delegate: Delegate) => void;
-}) {
-  const [displayedDelegates, setDisplayedDelegates] = useState(5);
-  return (
-    <>
-      <div
-        className={css({
-          fontSize: 16,
-          color: "contentAlt",
-        })}
-      >
-        {intro}
-      </div>
-
-      <div
-        className={css({
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          gap: 8,
-          paddingTop: 32,
-          paddingBottom: 24,
-          minHeight: 312,
-        })}
-      >
-        {delegates.slice(0, displayedDelegates).map((delegate) => {
-          return (
-            <DelegateBox
-              key={delegate.id}
-              delegate={delegate}
-              selectLabel="Choose"
-              onSelect={noop}
-            />
-          );
-        })}
-        {displayedDelegates < delegates.length && (
-          <ShadowBox>
-            <TextButton
-              label="Load more"
-              onClick={() => setDisplayedDelegates(displayedDelegates + 5)}
-              className={css({
-                width: "100%",
-                padding: "24px 0",
-                justifyContent: "center",
-              })}
-            />
-          </ShadowBox>
-        )}
-      </div>
-    </>
-  );
-}
-
-const MiniChart = memo(function MiniChart({
-  size = "small",
-}: {
-  size?: "small" | "medium";
-}) {
-  return (
-    size === "medium"
-      ? (
-        <svg width="45" height="18" fill="none">
-          <path
-            stroke="#9EA2B8"
-            strokeLinecap="round"
-            strokeWidth="1.5"
-            d="m1 10.607 1.66 2.61a2 2 0 0 0 1.688.926H7.31c.29 0 .576.063.84.185l4.684 2.168a2 2 0 0 0 2.142-.296l2.995-2.568a2 2 0 0 0 .662-1.137l1.787-9.191a2 2 0 0 1 .318-.756l.096-.138a2 2 0 0 1 3.303.02l1.421 2.107a2 2 0 0 1 .278.616l1.295 4.996a2 2 0 0 0 .431.815l1.691 1.932c.163.187.29.402.375.635l.731 2.015a2 2 0 0 0 3.029.955l.079-.056a2 2 0 0 0 .744-.99l2.539-7.42 2.477-5.005a2 2 0 0 1 2.924-.762L44 3.536"
-          />
-        </svg>
-      )
-      : (
-        <svg width="28" height="16" fill="none">
-          <path
-            stroke="#9EA2B8"
-            strokeLinecap="round"
-            strokeWidth="1.5"
-            d="m2 8.893.618 1.009a2 2 0 0 0 1.706.955h.83a2 2 0 0 1 .867.198l1.731.832a2 2 0 0 0 2.197-.309l.901-.802a2 2 0 0 0 .636-1.126l.902-4.819c.028-.148.085-.288.169-.414v0a1.119 1.119 0 0 1 1.87.011l.64.986c.113.175.197.368.248.571l.613 2.458a2 2 0 0 0 .411.804l.686.814c.145.172.257.37.332.582l.339.97a1.09 1.09 0 0 0 1.671.522v0a1.09 1.09 0 0 0 .394-.54l1.361-4.13.847-1.778a2 2 0 0 1 2.966-.77l.065.047"
-          />
-        </svg>
-      )
-  );
-});
-
-function ShadowBox({ children }: { children: ReactNode }) {
-  return (
-    <div
-      className={css({
-        width: "100%",
-        background: "background",
-        borderWidth: "1px 1px 0",
-        borderStyle: "solid",
-        borderColor: "gray:50",
-        boxShadow: `
-          0 2px 2px rgba(0, 0, 0, 0.1),
-          0 4px 10px rgba(18, 27, 68, 0.05),
-          inset 0 -1px 4px rgba(0, 0, 0, 0.05)
-        `,
-        borderRadius: 8,
-      })}
-    >
-      {children}
-    </div>
-  );
-}
-
-function DelegateBox({
-  delegate,
-  onSelect,
-  selectLabel = "Select",
-}: {
-  delegate: Delegate;
-  onSelect: (delegate: Delegate) => void;
-  selectLabel: string;
-}) {
-  const delegationRisk = getRedemptionRisk(delegate.interestRate);
-  return (
-    <ShadowBox key={delegate.id}>
-      <section
-        key={delegate.name}
-        className={css({
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          padding: "8px 16px",
-        })}
-      >
-        <div
-          className={css({
-            display: "flex",
-            flexDirection: "column",
-            width: "100%",
-            paddingBottom: 12,
-            borderBottom: "1px solid token(colors.borderSoft)",
-          })}
-        >
-          <div
-            className={css({
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              width: "100%",
-              fontSize: 20,
-              fontWeight: 500,
-              userSelect: "none",
-            })}
-          >
-            <h1 title={`${delegate.name} (${delegate.address})`}>
-              {delegate.name}
-            </h1>
-            <div
-              className={css({
-                display: "flex",
-                gap: 6,
-                alignItems: "center",
-              })}
-            >
-              <MiniChart />
-              {fmtnum(delegate.interestRate, "pct1z")}%
-            </div>
-          </div>
-          <div
-            className={css({
-              display: "flex",
-              justifyContent: "space-between",
-              width: "100%",
-              fontSize: 14,
-              color: "content",
-            })}
-          >
-            <div
-              className={css({
-                display: "flex",
-                gap: 8,
-                alignItems: "center",
-              })}
-            >
-              <Amount
-                value={delegate.boldAmount}
-                format="compact"
-                suffix=" BOLD"
-              />
-            </div>
-            <div
-              className={css({
-                display: "flex",
-                gap: 8,
-                alignItems: "center",
-              })}
-            >
-              <StatusDot mode={riskLevelToStatusMode(delegationRisk)} />
-              {formatRedemptionRisk(delegationRisk)}
-            </div>
-          </div>
-        </div>
-        <div
-          className={css({
-            display: "flex",
-            flexDirection: "column",
-            width: "100%",
-            paddingTop: 12,
-            fontSize: 14,
-            paddingBottom: 12,
-            borderBottom: "1px solid token(colors.borderSoft)",
-          })}
-        >
-          <div
-            className={css({
-              display: "flex",
-              justifyContent: "space-between",
-              width: "100%",
-              fontSize: 14,
-              color: "content",
-            })}
-          >
-            <div>Interest rate range</div>
-            <div>
-              {fmtnum(delegate.interestRateChange[0], "pct2")}
-              <span>-</span>
-              {fmtnum(delegate.interestRateChange[1], "pct2")}%
-            </div>
-          </div>
-          {delegate.fee && (
-            <div
-              className={css({
-                display: "flex",
-                justifyContent: "space-between",
-                width: "100%",
-                fontSize: 14,
-                color: "content",
-              })}
-            >
-              <div>
-                Fees <abbr title="per annum">p.a.</abbr>
-              </div>
-              <div title={`${fmtnum(delegate.fee, "pctfull")}%`}>
-                {fmtnum(delegate.fee, { digits: 4, scale: 100 })}%
-              </div>
-            </div>
-          )}
-        </div>
-        <div
-          className={css({
-            display: "flex",
-            justifyContent: "space-between",
-            width: "100%",
-            paddingTop: 16,
-            paddingBottom: 8,
-            fontSize: 14,
-          })}
-        >
-          <div
-            className={css({
-              display: "flex",
-              gap: 8,
-            })}
-          >
-            <TextButton
-              label={
-                <>
-                  Copy address
-                  <IconCopy size={16} />
-                </>
-              }
-              className={css({
-                fontSize: 14,
-              })}
-            />
-          </div>
-          <div>
-            <Button
-              label={selectLabel}
-              mode="primary"
-              size="small"
-              onClick={() => {
-                onSelect(delegate);
-              }}
-            />
-          </div>
-        </div>
-      </section>
-    </ShadowBox>
-  );
-}
