+++ frontend/app/src/comps/InterestRateField/MiniChart.tsx
@@ -1,0 +1,31 @@
+import { memo } from "react";
+
+export const MiniChart = memo(function MiniChart({
+  size = "small",
+}: {
+  size?: "small" | "medium";
+}) {
+  return (
+    size === "medium"
+      ? (
+        <svg width="45" height="18" fill="none">
+          <path
+            stroke="#9EA2B8"
+            strokeLinecap="round"
+            strokeWidth="1.5"
+            d="m1 10.607 1.66 2.61a2 2 0 0 0 1.688.926H7.31c.29 0 .576.063.84.185l4.684 2.168a2 2 0 0 0 2.142-.296l2.995-2.568a2 2 0 0 0 .662-1.137l1.787-9.191a2 2 0 0 1 .318-.756l.096-.138a2 2 0 0 1 3.303.02l1.421 2.107a2 2 0 0 1 .278.616l1.295 4.996a2 2 0 0 0 .431.815l1.691 1.932c.163.187.29.402.375.635l.731 2.015a2 2 0 0 0 3.029.955l.079-.056a2 2 0 0 0 .744-.99l2.539-7.42 2.477-5.005a2 2 0 0 1 2.924-.762L44 3.536"
+          />
+        </svg>
+      )
+      : (
+        <svg width="28" height="16" fill="none">
+          <path
+            stroke="#9EA2B8"
+            strokeLinecap="round"
+            strokeWidth="1.5"
+            d="m2 8.893.618 1.009a2 2 0 0 0 1.706.955h.83a2 2 0 0 1 .867.198l1.731.832a2 2 0 0 0 2.197-.309l.901-.802a2 2 0 0 0 .636-1.126l.902-4.819c.028-.148.085-.288.169-.414v0a1.119 1.119 0 0 1 1.87.011l.64.986c.113.175.197.368.248.571l.613 2.458a2 2 0 0 0 .411.804l.686.814c.145.172.257.37.332.582l.339.97a1.09 1.09 0 0 0 1.671.522v0a1.09 1.09 0 0 0 .394-.54l1.361-4.13.847-1.778a2 2 0 0 1 2.966-.77l.065.047"
+          />
+        </svg>
+      )
+  );
+});
